fix: add validation ketika picture masih null

+++ src/controllers/profile.js
@@ -66,7 +66,7 @@ exports.updateProfile = async (req, res) => {
 
     const user = await User.findByPk(id);
 
-    if (req.files) {
+    if (req.files && user.picture) {
       deleteFile(cloudPathToFileName(user.picture));
       req.files.forEach(async (pic) => {
         data.picture = pic.path;
