+++ src/controllers/user.js
@@ -170,7 +170,7 @@ exports.updateUser = async (req, res) => {
       return responseHandler(res, 400, error);
     }
 
-    if (req.files) {
+    if (req.files && user.picture) {
       deleteFile(cloudPathToFileName(user.picture));
       req.files.forEach(async (pic) => {
         data.picture = pic.path;
